only norminette and garbage file left to push

+++ libft/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: hfeufeu <hfeufeu@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/17 16:08:16 by hfeufeu           #+#    #+#             */
-/*   Updated: 2025/02/17 13:29:36 by hfeufeu          ###   ########.fr       */
+/*   Updated: 2025/02/18 11:55:38 by hfeufeu          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
-	int	n;
-	int	sign;
+	long	n;
+	int		sign;
 
 	n = 0;
 	sign = 1;
@@ -29,7 +29,5 @@ int	ft_atoi(const char *nptr)
 	}
 	while (*nptr >= '0' && *nptr <= '9')
 		n = (n * 10) + (*nptr++ - '0');
-	// if (n == 2147483648)
-	// 	err_handle(ERR_ARG);
 	return (n * sign);
 }
